+++ egui_glium/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Changed
+
+* FileStorage::from_path now takes `Into<Path>` instead of `String`
+
 ## 0.4.0 - 2020-11-28
 
 Started changelog. Features: